Sonar fix
